create exec_const
